requirements.txt
absl-py==2.1.0
altair==5.5.0
astunparse==1.6.3
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
deepface==0.0.93
filelock==3.16.1
fire==0.7.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.12.23
fonttools==4.55.3
gast==0.6.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.69.0
gunicorn==23.0.0
h5py==3.12.1
idna==3.10
itsdangerous==2.2.0
jax==0.5.0
jaxlib==0.5.0
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.8.0
kiwisolver==1.4.8
libclang==18.1.1
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mediapipe==0.10.20
ml-dtypes==0.4.1
mtcnn==1.0.0
namex==0.0.8
narwhals==1.23.0
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==4.25.5
pyarrow==19.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.36.1
requests==2.32.3
retina-face==0.0.17
rich==13.9.4
rpds-py==0.22.3
scipy==1.15.1
sentencepiece==0.2.0
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
sounddevice==0.5.1
soupsieve==2.6
streamlit==1.41.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
tf_keras==2.18.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
